metodo gravar adicionado na marca e modelo

+++ src/java/model/Marca.java
@@ -14,19 +14,19 @@ import java.util.List;
  * @author Raphael
  */
 public class Marca {
-    private long id;
+    private int id;
     private String nome;
 
-    public Marca(long id, String nome) {
+    public Marca(int id, String nome) {
         this.id = id;
         this.nome = nome;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
     
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
